move diagnostics classes to dedicated namespace and file
